part4 complete - last3 optional exercises not done

+++ part4/blog/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog');
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username : 1, name : 1});
@@ -25,13 +17,14 @@ blogRouter.post('/', async (request, response) => {
 	}
 
   const body = request.body
-  const token = getTokenFrom(request)
+  console.log(request.token)
+  const token = request.token
   const decodedToken = jwt.verify(token, process.env.SECRET)
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
-
+  console.log(user)
   const blog = new Blog({
     title : body.title,
     author : body.author,
@@ -49,8 +42,17 @@ blogRouter.post('/', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async(request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = Blog.findById(request.params.id)
+  console.log(blog)
+  
   const result = await Blog.findByIdAndRemove(request.params.id);
-  console.log(result);
+  console.log(result)
   response.status(204).end();
 })
